Deepen Kafka delivery contract
