+++ packages/runtime/src/kafka-delivery-contract.test.ts
@@ -1,15 +1,19 @@
 import { describe, expect, it } from "@effect/vitest";
 import { defineViewServerConfig } from "@effect-view-server/config";
 import { Buffer } from "node:buffer";
-import { Schema } from "effect";
+import { Effect, Exit, Schema } from "effect";
 import {
+  publishAndCommitKafkaDecodedBatch,
   planKafkaDecodedPublishRuns,
   planKafkaUpsertPublishRuns,
+  recordAndCommitKeylessKafkaMessage,
   type DecodedKafkaBatchMessage,
   type DecodedKafkaBatchTombstoneMessage,
   type DecodedKafkaBatchUpsertMessage,
+  type KafkaDeliveryRuntimeClient,
   type KafkaConsumerMessage,
 } from "./kafka-delivery-contract";
+import { makeViewServerKafkaHealthLedger } from "./kafka-health";
 
 const Order = Schema.Struct({
   id: Schema.String,
@@ -37,10 +41,11 @@ const viewServer = defineViewServerConfig({
 type Topics = typeof viewServer.topics;
 
 const kafkaMessage = (input: {
+  readonly commit?: KafkaConsumerMessage["commit"];
   readonly offset: bigint;
   readonly sourceTopic: string;
 }): KafkaConsumerMessage => ({
-  commit: () => undefined,
+  commit: input.commit ?? (() => undefined),
   headers: new Map(),
   key: Buffer.from(`key-${input.offset}`),
   metadata: {},
@@ -62,6 +67,7 @@ const kafkaMessage = (input: {
 });
 
 const upsertMessage = (input: {
+  readonly commit?: KafkaConsumerMessage["commit"];
   readonly offset: bigint;
   readonly row: object;
   readonly rowKey: string;
@@ -74,6 +80,7 @@ const upsertMessage = (input: {
     viewServerTopic: input.viewServerTopic,
   },
   message: kafkaMessage({
+    ...(input.commit === undefined ? {} : { commit: input.commit }),
     offset: input.offset,
     sourceTopic: input.sourceTopic,
   }),
@@ -83,6 +90,7 @@ const upsertMessage = (input: {
 });
 
 const tombstoneMessage = (input: {
+  readonly commit?: KafkaConsumerMessage["commit"];
   readonly offset: bigint;
   readonly rowKey: string;
   readonly sourceTopic: string;
@@ -94,6 +102,7 @@ const tombstoneMessage = (input: {
     viewServerTopic: input.viewServerTopic,
   },
   message: kafkaMessage({
+    ...(input.commit === undefined ? {} : { commit: input.commit }),
     offset: input.offset,
     sourceTopic: input.sourceTopic,
   }),
@@ -101,6 +110,54 @@ const tombstoneMessage = (input: {
   nowMillis: Number(input.offset),
   sourceTopic: input.sourceTopic,
 });
+
+const makeDeliveryClient = (
+  operations: Array<string>,
+  options?: {
+    readonly publishFails?: boolean;
+    readonly publishFailsForTopic?: Extract<keyof Topics, string>;
+  },
+): KafkaDeliveryRuntimeClient<Topics> => ({
+  delete: (topic, key) =>
+    Effect.sync(() => {
+      operations.push(`delete:${topic}:${key}`);
+    }),
+  publishManyDecodedRowsWithStorageKeys: (topic, rows) =>
+    options?.publishFails === true || options?.publishFailsForTopic === topic
+      ? Effect.sync(() => {
+          operations.push(`publish-fail:${topic}:${rows.map((row) => row.storageKey).join(",")}`);
+        }).pipe(Effect.andThen(Effect.die("publish-down")))
+      : Effect.sync(() => {
+          operations.push(`publish:${topic}:${rows.map((row) => row.storageKey).join(",")}`);
+        }),
+});
+
+const makeKafkaHealthLedger = () =>
+  makeViewServerKafkaHealthLedger<Topics>({
+    regions: {
+      local: "localhost:9092",
+    },
+    startFrom: {
+      consumerGroupId: "view-server-delivery-contract-test",
+      fallbackMode: "earliest",
+      mode: "earliest",
+    },
+    topics: {
+      "orders-source": {
+        regions: ["local"],
+        viewServerTopic: "orders",
+      },
+      "trades-source": {
+        regions: ["local"],
+        viewServerTopic: "trades",
+      },
+    },
+  });
+
+const healthRefresh = (operations: Array<string>) =>
+  Effect.sync(() => {
+    operations.push("health");
+  });
 
 describe("Kafka delivery contract", () => {
   it("plans contiguous upsert runs and tombstone runs without reordering messages", () => {
@@ -287,4 +344,166 @@ describe("Kafka delivery contract", () => {
       },
     ]);
   });
+
+  it.effect("publishes and commits decoded runs in Kafka order", () =>
+    Effect.gen(function* () {
+      const operations: Array<string> = [];
+      const firstUpsert = upsertMessage({
+        commit: () => {
+          operations.push("commit:1");
+        },
+        offset: 1n,
+        row: { id: "a", price: 10 },
+        rowKey: "a",
+        sourceTopic: "orders-source",
+        viewServerTopic: "orders",
+      });
+      const secondUpsert = upsertMessage({
+        commit: () => {
+          operations.push("commit:2");
+        },
+        offset: 2n,
+        row: { id: "b", price: 20 },
+        rowKey: "b",
+        sourceTopic: "orders-source",
+        viewServerTopic: "orders",
+      });
+      const tombstone = tombstoneMessage({
+        commit: () => {
+          operations.push("commit:3");
+        },
+        offset: 3n,
+        rowKey: "a",
+        sourceTopic: "orders-source",
+        viewServerTopic: "orders",
+      });
+      const finalUpsert = upsertMessage({
+        commit: () => {
+          operations.push("commit:4");
+        },
+        offset: 4n,
+        row: { id: "c", price: 30 },
+        rowKey: "c",
+        sourceTopic: "orders-source",
+        viewServerTopic: "orders",
+      });
+
+      yield* publishAndCommitKafkaDecodedBatch(
+        makeDeliveryClient(operations),
+        healthRefresh(operations),
+        makeKafkaHealthLedger(),
+        "local",
+        [firstUpsert, secondUpsert, tombstone, finalUpsert],
+      );
+
+      expect(operations).toStrictEqual([
+        "publish:orders:a,b",
+        "commit:1",
+        "commit:2",
+        "health",
+        "delete:orders:a",
+        "commit:3",
+        "health",
+        "publish:orders:c",
+        "commit:4",
+        "health",
+      ]);
+    }),
+  );
+
+  it.effect("does not commit decoded messages after publish failure", () =>
+    Effect.gen(function* () {
+      const operations: Array<string> = [];
+      const message = upsertMessage({
+        commit: () => {
+          operations.push("commit:1");
+        },
+        offset: 1n,
+        row: { id: "a", price: 10 },
+        rowKey: "a",
+        sourceTopic: "orders-source",
+        viewServerTopic: "orders",
+      });
+
+      const exit = yield* Effect.exit(
+        publishAndCommitKafkaDecodedBatch(
+          makeDeliveryClient(operations, { publishFails: true }),
+          healthRefresh(operations),
+          makeKafkaHealthLedger(),
+          "local",
+          [message],
+        ),
+      );
+
+      expect(Exit.isFailure(exit)).toBe(true);
+      expect(operations).toStrictEqual(["publish-fail:orders:a", "health"]);
+    }),
+  );
+
+  it.effect("commits each successful compatible upsert run before publishing the next run", () =>
+    Effect.gen(function* () {
+      const operations: Array<string> = [];
+      const ordersMessage = upsertMessage({
+        commit: () => {
+          operations.push("commit:orders");
+        },
+        offset: 1n,
+        row: { id: "a", price: 10 },
+        rowKey: "a",
+        sourceTopic: "orders-source",
+        viewServerTopic: "orders",
+      });
+      const tradesMessage = upsertMessage({
+        commit: () => {
+          operations.push("commit:trades");
+        },
+        offset: 2n,
+        row: { id: "t1", quantity: 20 },
+        rowKey: "t1",
+        sourceTopic: "trades-source",
+        viewServerTopic: "trades",
+      });
+
+      const exit = yield* Effect.exit(
+        publishAndCommitKafkaDecodedBatch(
+          makeDeliveryClient(operations, { publishFailsForTopic: "trades" }),
+          healthRefresh(operations),
+          makeKafkaHealthLedger(),
+          "local",
+          [ordersMessage, tradesMessage],
+        ),
+      );
+
+      expect(Exit.isFailure(exit)).toBe(true);
+      expect(operations).toStrictEqual([
+        "publish:orders:a",
+        "commit:orders",
+        "health",
+        "publish-fail:trades:t1",
+        "health",
+      ]);
+    }),
+  );
+
+  it.effect("records and commits keyless source messages as skipped delivery failures", () =>
+    Effect.gen(function* () {
+      const operations: Array<string> = [];
+      const message = kafkaMessage({
+        commit: () => {
+          operations.push("commit:5");
+        },
+        offset: 5n,
+        sourceTopic: "orders-source",
+      });
+
+      yield* recordAndCommitKeylessKafkaMessage(
+        healthRefresh(operations),
+        makeKafkaHealthLedger(),
+        "local",
+        message,
+      );
+
+      expect(operations).toStrictEqual(["health", "commit:5", "health"]);
+    }),
+  );
 });
